1l - one cast too many
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@17385 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/cache2.c
@@ -289,7 +289,7 @@ int stream_enable_cache(stream_t *stream,int size,int min,int seek_limit){
 #endif
     // wait until cache is filled at least prefill_init %
     mp_msg(MSGT_CACHE,MSGL_V,"CACHE_PRE_INIT: %"PRId64" [%"PRId64"] %"PRId64"  pre:%d  eof:%d  \n",
-	(int64_t)s->min_filepos,(int64_t)s->read_filepos,(int64_t)s->max_filepos,min,(int64_t)s->eof);
+	(int64_t)s->min_filepos,(int64_t)s->read_filepos,(int64_t)s->max_filepos,min,s->eof);
     while(s->read_filepos<s->min_filepos || s->max_filepos-s->read_filepos<min){
 	mp_msg(MSGT_CACHE,MSGL_STATUS,MSGTR_CacheFill,
 	    100.0*(float)(s->max_filepos-s->read_filepos)/(float)(s->buffer_size),
